+++ config.php
@@ -11,7 +11,6 @@ use humhub\compat\HForm;
 use humhub\modules\authKeycloak\Events;
 use humhub\modules\user\authclient\Collection;
 use humhub\modules\user\controllers\AuthController;
-use humhub\modules\user\controllers\RegistrationController;
 use humhub\modules\user\models\Auth;
 use humhub\modules\user\models\Group;
 use humhub\modules\user\models\GroupUser;
@@ -27,18 +26,8 @@ return [
     'events' => [
         [
             'class' => AuthController::class,
-            'event' => AuthController::EVENT_BEFORE_ACTION,
-            'callback' => [Events::class, 'onUserAuthControllerBeforeAction']
-        ],
-        [
-            'class' => AuthController::class,
             'event' => AuthController::EVENT_AFTER_LOGIN,
             'callback' => [Events::class, 'onAfterLogin']
-        ],
-        [
-            'class' => RegistrationController::class,
-            'event' => RegistrationController::EVENT_BEFORE_ACTION,
-            'callback' => [Events::class, 'onUserRegistrationControllerBeforeAction']
         ],
         [
             'class' => HForm::class,